rebuild 'configure' by the 'bootstrap' script
The bootstrap script is normally used once while creating the source
tarball and i screating in the end the configure script. But we need to
create a freshly generated configure script and other autoconf related
files to get newer platforms working like arm64 (see also 64ad53)
To get the 'bootstrap' script usable we need the autotools at all, For
this adding dh-autoreconf to the B-D.

Calling the bootstrap script right before the configure call to get a
actual configure script that supports all platforms.