debian/rules: adjust the targets after dh version change
Especially after the move to dh10 the autotools need a file
config.rpath. But for libvmime we can ignore this file and just touching
this file so autoreconf will be happy.

Because of dh_autoreconf we don't need to call bootstrap then again,
which does the same basically as dh_autoreconf. We can remove that call.