debian/rules: modify override targets to new vmime requirements
Due the CMake system used by vmime and also the way vmime needs to be
configured within a Unix system the target 'dh_auto_configure' and
'dh_auto_build' needs to be changed in a different way than the
sequenzer default is.

As we want to build out of tree we preapre also some needed variables
before the build can start.